requirements.txt

Django==5.0.4
djangorestframework==3.15.1
drf_yasg==1.21.7
